Conflicts:
	assets/sass/modules/dashboard-panel/_module_style.scss
	pom.xml
	src/main/java/de/maredit/tar/configs/ApplicationConfig.java
	src/main/java/de/maredit/tar/controllers/VacationController.java
	src/main/java/de/maredit/tar/properties/VacationProperties.java
	src/main/resources/static/css/style.css.map
	src/main/resources/vacationConfig.yaml

+++ src/main/java/de/maredit/tar/properties/VacationProperties.java
@@ -11,14 +11,14 @@ import de.maredit.tar.models.UserHoliday;
 @ConfigurationProperties(locations = "classpath:vacationConfig.yaml")
 public class VacationProperties {
 
-  private int defaultVacationDays;
+  private double defaultVacationDays;
   private List<UserHoliday> userHolidays;
 
-  public int getDefaultVacationDays() {
+  public double getDefaultVacationDays() {
     return defaultVacationDays;
   }
 
-  public void setDefaultVacationDays(int defaultVacationDays) {
+  public void setDefaultVacationDays(double defaultVacationDays) {
     this.defaultVacationDays = defaultVacationDays;
   }
